+++ src/main/java/net/mohron/skyclaims/PluginInfo.java
@@ -25,7 +25,7 @@ public final class PluginInfo {
     public static final String VERSION = "@VERSION@";
     public static final String DESCRIPTION = "@DESCRIPTION@";
     public static final String AUTHORS = "Mohron, Cossacksman";
-    public static final String GP_VERSION = "4.0.0.332";
+    public static final String GP_VERSION = "4.0.0.378";
     public static final double GP_API_VERSION = 0.6;
-    public static final String NUCLEUS_VERSION = "1.0.1";
+    public static final String NUCLEUS_VERSION = "1.1.0";
 }