+++ src/main/java/net/mohron/skyclaims/command/CommandBase.java
@@ -18,11 +18,61 @@
 
 package net.mohron.skyclaims.command;
 
+import com.google.common.collect.Lists;
 import net.mohron.skyclaims.SkyClaims;
+import net.mohron.skyclaims.world.Island;
+import org.spongepowered.api.command.CommandException;
+import org.spongepowered.api.command.CommandResult;
+import org.spongepowered.api.command.CommandSource;
+import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.spec.CommandExecutor;
+import org.spongepowered.api.entity.living.player.Player;
+import org.spongepowered.api.text.Text;
+import org.spongepowered.api.text.format.TextColors;
+
+import java.util.List;
+import java.util.UUID;
+
+import javax.annotation.Nonnull;
 
 public abstract class CommandBase implements CommandExecutor {
 
     protected static final SkyClaims PLUGIN = SkyClaims.getInstance();
+
+    public static abstract class IslandCommand extends CommandBase implements CommandRequirement.RequiresIsland {
+
+        protected static final Text ISLAND = Text.of("island");
+
+        @Nonnull @Override public CommandResult execute(@Nonnull CommandSource src, @Nonnull CommandContext args) throws CommandException {
+            if (!(src instanceof Player)) {
+                throw new CommandException(Text.of(TextColors.RED, "This command can only be used by a player!"));
+            }
+            if (!args.hasAny(ISLAND)) {
+                return execute(
+                    (Player) src,
+                    Island.get(((Player) src).getLocation())
+                        .orElseThrow(() -> new CommandException(Text.of(TextColors.RED, "You must be on an island to use this command."))),
+                    args);
+            } else {
+                List<Island> islands = Lists.newArrayList();
+                args.<UUID>getAll(ISLAND).forEach(i -> Island.get(i).ifPresent(islands::add));
+                if (islands.size() == 1) {
+                    return execute((Player) src, islands.get(0), args);
+                } else {
+                    throw new CommandException(Text.of(TextColors.RED, "Multiple island support not yet implemented!"));
+                }
+            }
+        }
+    }
+
+    public static abstract class PlayerCommand extends CommandBase implements CommandRequirement.RequiresPlayer {
+
+        @Nonnull @Override public CommandResult execute(@Nonnull CommandSource src, @Nonnull CommandContext args) throws CommandException {
+            if (src instanceof Player) {
+                return execute((Player) src, args);
+            }
+            throw new CommandException(Text.of(TextColors.RED, "This command can only be used by a player!"));
+        }
+    }
 
 }